project/backoffice-api/requirements.txt
aiosqlite==0.16.1
appdirs==1.4.4
asgiref==3.3.4
asyncpg==0.23.0
attrs==21.2.0
bcrypt==3.2.0
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coverage==5.5
cryptography==3.4.7
ecdsa==0.17.0
Faker==8.8.1
fastapi==0.65.2
flake8==3.9.2
gunicorn==20.1.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
iso8601==0.1.14
isort==5.9.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
passlib[bcrypt]==1.7.4
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-cov==2.12.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0